+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## 1.1.0 - 2026-08-05
+- New: Checkout Session, Subscription, and Billing Portal Session models.
+- Fix: verify account contract
 
 ## 1.0.0